+++ packages/editor/web/src/RichContentEditor/I18nRichContentEditor.tsx
@@ -21,6 +21,8 @@ export default class I18nRichContentEditor extends Component<Partial<RichContent
 
   getT = () => this.editor.getT();
 
+  getEditorContainer = () => this.editor.getEditorContainer();
+
   getPlugins = () => this.editor.getPlugins();
 
   focus = () => this.editor.focus();