+++ packages/plugin-link/web/src/tiptap.ts
@@ -123,24 +123,16 @@ export const createTiptapExtensions: CreateRicosExtensions = defaultOptions => [
 
       addPasteRules() {
         return [
-          markPasteRule({
-            find: urlRegex,
-            type: this.type,
-            getAttributes: (match: RegExpExecArray) => ({
-              link: { url: match[0] },
-            }),
-          }),
+          markPasteRule(urlRegex, this.type, (match: RegExpExecArray) => ({
+            link: { url: match[0] },
+          })),
         ];
       },
 
       addInputRules() {
         return [
-          markInputRule({
-            find: urlRegexExactWithSpace,
-            type: this.type,
-            getAttributes: (match: RegExpExecArray) => {
-              return { link: { url: match[0].trim() } };
-            },
+          markInputRule(urlRegexExactWithSpace, this.type, (match: RegExpExecArray) => {
+            return { link: { url: match[0].trim() } };
           }),
         ];
       },