+++ packages/ricos-editor/web/src/RicosEditor.tsx
@@ -403,6 +403,7 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
       linkSettings,
       ...biFunctions,
       experiments,
+      editorContainer: this.editor?.getEditorContainer() as HTMLElement,
     };
     const baseStyles = { flex: 'none', webkitTapHighlightColor: 'transparent' };
     const baseMobileStyles = { ...baseStyles, position: 'sticky', top: 0, zIndex: 9 };
@@ -503,12 +504,10 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
     if (!tiptapEditorModule) {
       return null;
     }
-    const { RicosTiptapEditor, RichContentAdapter } = tiptapEditorModule;
+    const { RicosTiptapEditor, RichContentAdapter, draftToTiptap } = tiptapEditorModule;
     const { content, injectedContent, plugins } = this.props;
     const { tiptapToolbar } = this.state;
-    const initialContent = tiptapEditorModule.draftToTiptap(
-      content ?? injectedContent ?? emptyDraftContent
-    );
+    const initialContent = draftToTiptap(content ?? injectedContent ?? emptyDraftContent);
     const { localeData } = this.state;
     const { locale, localeResource } = localeData;
     const extensions =