+++ packages/tiptap-editor/web/src/core/extensions/commands.ts
@@ -1,6 +1,6 @@
 import { Extension } from '@tiptap/core';
 import * as insertNode from '../commands/insertNode';
-import * as updateNodeAttrsById from '../commands/updateNodeAttrsById';
+import * as updateNodeById from '../commands/updateNodeById';
 
 export const Commands = Extension.create({
   name: 'commands',
@@ -8,7 +8,7 @@ export const Commands = Extension.create({
   addCommands() {
     return {
       ...insertNode,
-      ...updateNodeAttrsById,
+      ...updateNodeById,
     };
   },
 });