+++ packages/toolbars-new/web/src/Toolbar/buttonsListCreatorConsts.tsx
@@ -94,6 +94,7 @@ export const alignmentsModalData = [
     },
     commandKey: 'left',
     icon: AlignLeftIcon,
+    dataHook: 'textAlignmentButton_left',
   },
   {
     text: 'AlignTextCenterButton_Tooltip',
@@ -104,6 +105,7 @@ export const alignmentsModalData = [
     },
     commandKey: 'center',
     icon: AlignTextCenterIcon,
+    dataHook: 'textAlignmentButton_center',
   },
   {
     text: 'AlignTextRightButton_Tooltip',
@@ -114,6 +116,7 @@ export const alignmentsModalData = [
     },
     commandKey: 'right',
     icon: AlignRightIcon,
+    dataHook: 'textAlignmentButton_right',
   },
   {
     text: 'AlignTextJustifyButton_Tooltip',
@@ -124,6 +127,7 @@ export const alignmentsModalData = [
     },
     commandKey: 'justify',
     icon: AlignJustifyIcon,
+    dataHook: 'textAlignmentButton_justify',
   },
 ];
 
@@ -377,7 +381,7 @@ export const buttonsFullData: Record<string, buttonsFullDataType> = {
   },
   Bold: {
     icon: BoldIcon,
-    dataHook: 'textInlineStyleButton_BOLD',
+    dataHook: 'textInlineStyleButton_Bold',
     tooltip: 'BoldButton_Tooltip',
     tooltipShortcut: {
       MacOS: ' (⌘B)',
@@ -387,7 +391,7 @@ export const buttonsFullData: Record<string, buttonsFullDataType> = {
   },
   Italic: {
     icon: ItalicIcon,
-    dataHook: 'textInlineStyleButton_ITALIC',
+    dataHook: 'textInlineStyleButton_Italic',
     tooltip: 'ItalicButton_Tooltip',
     tooltipShortcut: {
       MacOS: ' (⌘I)',
@@ -397,7 +401,7 @@ export const buttonsFullData: Record<string, buttonsFullDataType> = {
   },
   Underline: {
     icon: UnderlineIcon,
-    dataHook: 'textInlineStyleButton_UNDERLINE',
+    dataHook: 'textInlineStyleButton_Underline',
     tooltip: 'UnderlineButton_Tooltip',
     tooltipShortcut: {
       MacOS: ' (⌘U)',
@@ -455,7 +459,7 @@ export const buttonsFullData: Record<string, buttonsFullDataType> = {
     type: 'button',
   },
   Alignment: {
-    dataHook: 'Alignment',
+    dataHook: 'textDropDownButton_Alignment',
     arrow: true,
     tooltip: 'AlignTextDropdownButton_Tooltip',
     type: 'modal',